requirements.txt
Appium-Python-Client==2.5.2
pytest==7.2.0
pytest-html==3.2.0
selenium==4.9.0
